refactor scheme file reduce duplicate code by scheme inheritance
