+++ backend/routers/users.py
@@ -3,11 +3,11 @@ from typing import List
 from fastapi import Depends, HTTPException, APIRouter
 from sqlalchemy.orm import Session
 
-import schemas
 import crud
 from auth import oauth2
 from auth.oauth2 import oauth2_schema
 from database import get_db
+from schemas.user import User, UserRegister, UserDetails, ProfileUpdate, UserUpdate
 
 router = APIRouter(
     prefix='/users',
@@ -15,14 +15,14 @@ router = APIRouter(
 )
 
 
-@router.get('', response_model=List[schemas.UserDisplay])
+@router.get('', response_model=List[UserDetails])
 def get_all_users(db: Session = Depends(get_db)):
     return crud.get_all_users(db)
 
 
 @router.post('')
 def register_user(
-        user: schemas.UserRegister, db: Session = Depends(get_db)
+        user: UserRegister, db: Session = Depends(get_db)
 ):
     new_user = crud.create_user(db=db, user=user)
     access_token = oauth2.create_access_token(data={'sub': user.email})
@@ -37,10 +37,10 @@ def register_user(
     }
 
 
-@router.get('/profile', response_model=schemas.UserDetails)
+@router.get('/profile', response_model=UserDetails)
 def get_user(
         db: Session = Depends(get_db),
-        current_user: schemas.User = Depends(oauth2.get_current_user)
+        current_user: User = Depends(oauth2.get_current_user)
 ):
     db_user = crud.get_user(db, current_user.id)
     if db_user is None:
@@ -54,9 +54,9 @@ def get_user(
 
 
 @router.patch("/profile")
-async def update_user(user: schemas.ProfileUpdate,
+async def update_user(user: ProfileUpdate,
                       db: Session = Depends(get_db),
-                      current_user: schemas.User = Depends(oauth2.get_current_user),
+                      current_user: User = Depends(oauth2.get_current_user),
                       token: str = Depends(oauth2_schema)):
     updated_user = crud.update_user(db, current_user.id, user)
     return {
@@ -70,9 +70,9 @@ async def update_user(user: schemas.ProfileUpdate,
 
 
 @router.patch("/{user_id}")
-async def update_user(user: schemas.UserUpdate,
+async def update_user(user: UserUpdate,
                       db: Session = Depends(get_db),
-                      current_user: schemas.User = Depends(oauth2.get_current_user)):
+                      current_user: User = Depends(oauth2.get_current_user)):
     updated_user = crud.update_user(db, user.id, user)
     return {
         'id': updated_user.id,
@@ -82,10 +82,10 @@ async def update_user(user: schemas.UserUpdate,
     }
 
 
-@router.get('/{user_id}', response_model=schemas.UserDisplay)
+@router.get('/{user_id}', response_model=UserDetails)
 def get_user(user_id: int,
              db: Session = Depends(get_db),
-             current_user: schemas.User = Depends(oauth2.get_current_user)
+             current_user: User = Depends(oauth2.get_current_user)
              ):
     db_user = crud.get_user(db, user_id)
     if db_user is None:
